fix regularization of holidays test
