+++ tests/test_regularization.py
@@ -61,7 +61,7 @@ def test_regularization_holidays():
 
     m = NeuralProphet(
         epochs=20,
-        batch_size=64,
+        batch_size=32,
         learning_rate=0.1,
         yearly_seasonality=False,
         weekly_seasonality=False,
@@ -80,8 +80,8 @@ def test_regularization_holidays():
             to_reduce.append(weight_list[0][0][0])
         else:
             to_preserve.append(weight_list[0][0][0])
-    # print(to_reduce)
-    # print(to_preserve)
+    print(f"To reduce (< 0.2) {to_reduce}")
+    print(f"To preserve (> 0.5) {to_preserve}")
     assert np.mean(to_reduce) < 0.2
     assert np.mean(to_preserve) > 0.5
