+++ tests/test_train_config.py
@@ -78,49 +78,3 @@ def test_custom_lr_scheduler():
         freq="D",
         scheduler="OneCycleLR",
     )
-
-
-# def test_continue_training_checkpoint():
-#     df = pd.read_csv(PEYTON_FILE, nrows=NROWS)
-#     m = NeuralProphet(
-#         epochs=EPOCHS,
-#         batch_size=BATCH_SIZE,
-#         learning_rate=LR,
-#         n_lags=6,
-#         n_forecasts=3,
-#         n_changepoints=0,
-#     )
-#     metrics = m.fit(df, checkpointing=True, freq="D")
-#     metrics2 = m.fit(df, freq="D", continue_training=True, epochs=ADDITIONAL_EPOCHS)
-#     assert metrics["Loss"].min() >= metrics2["Loss"].min()
-
-
-# def test_continue_training_with_scheduler_selection():
-#     df = pd.read_csv(PEYTON_FILE, nrows=NROWS)
-#     m = NeuralProphet(
-#         epochs=EPOCHS,
-#         batch_size=BATCH_SIZE,
-#         learning_rate=LR,
-#         n_lags=6,
-#         n_forecasts=3,
-#         n_changepoints=0,
-#     )
-#     metrics = m.fit(df, checkpointing=True, freq="D")
-#     # Continue training with StepLR
-#     metrics2 = m.fit(df, freq="D", continue_training=True, epochs=ADDITIONAL_EPOCHS, scheduler="StepLR")
-#     assert metrics["Loss"].min() >= metrics2["Loss"].min()
-
-
-# def test_save_load_continue_training():
-#     df = pd.read_csv(PEYTON_FILE, nrows=NROWS)
-#     m = NeuralProphet(
-#         epochs=EPOCHS,
-#         n_lags=6,
-#         n_forecasts=3,
-#         n_changepoints=0,
-#     )
-#     metrics = m.fit(df, checkpointing=True, freq="D")
-#     save(m, "test_model.pt")
-#     m2 = load("test_model.pt")
-#     metrics2 = m2.fit(df, continue_training=True, epochs=ADDITIONAL_EPOCHS, scheduler="StepLR")
-#     assert metrics["Loss"].min() >= metrics2["Loss"].min()
