+++ tests/test_utils.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python3
 
-import io
 import logging
 import os
 import pathlib
@@ -8,7 +7,7 @@ import pathlib
 import pandas as pd
 import pytest
 
-from neuralprophet import NeuralProphet, df_utils, load, save
+from neuralprophet import NeuralProphet, df_utils
 
 log = logging.getLogger("NP.test")
 log.setLevel("ERROR")
@@ -39,64 +38,3 @@ def test_create_dummy_datestamps():
     m = NeuralProphet(epochs=EPOCHS, batch_size=BATCH_SIZE, learning_rate=LR)
     _ = m.fit(df_dummy)
     _ = m.make_future_dataframe(df_dummy, periods=365, n_historic_predictions=True)
-
-
-def test_save_load():
-    df = pd.read_csv(PEYTON_FILE, nrows=NROWS)
-    m = NeuralProphet(
-        epochs=EPOCHS,
-        batch_size=BATCH_SIZE,
-        learning_rate=LR,
-        n_lags=6,
-        n_forecasts=3,
-        n_changepoints=0,
-    )
-    _ = m.fit(df, freq="D")
-    future = m.make_future_dataframe(df, periods=3)
-    forecast = m.predict(df=future)
-    log.info("testing: save")
-    save(m, "test_model.pt")
-
-    log.info("testing: load")
-    m2 = load("test_model.pt")
-    forecast2 = m2.predict(df=future)
-
-    m3 = load("test_model.pt", map_location="cpu")
-    forecast3 = m3.predict(df=future)
-
-    # Check that the forecasts are the same
-    pd.testing.assert_frame_equal(forecast, forecast2)
-    pd.testing.assert_frame_equal(forecast, forecast3)
-
-
-def test_save_load_io():
-    df = pd.read_csv(PEYTON_FILE, nrows=NROWS)
-    m = NeuralProphet(
-        epochs=EPOCHS,
-        batch_size=BATCH_SIZE,
-        learning_rate=LR,
-        n_lags=6,
-        n_forecasts=3,
-        n_changepoints=0,
-    )
-    _ = m.fit(df, freq="D")
-    future = m.make_future_dataframe(df, periods=3)
-    forecast = m.predict(df=future)
-
-    # Save the model to an in-memory buffer
-    log.info("testing: save to buffer")
-    buffer = io.BytesIO()
-    save(m, buffer)
-    buffer.seek(0)  # Reset buffer position to the beginning
-
-    log.info("testing: load from buffer")
-    m2 = load(buffer)
-    forecast2 = m2.predict(df=future)
-
-    buffer.seek(0)  # Reset buffer position to the beginning for another load
-    m3 = load(buffer, map_location="cpu")
-    forecast3 = m3.predict(df=future)
-
-    # Check that the forecasts are the same
-    pd.testing.assert_frame_equal(forecast, forecast2)
-    pd.testing.assert_frame_equal(forecast, forecast3)
